+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 * Fix #1285: removed references to manually calling registerCustomKind
 * Fix #3334: adding basic support for server side apply.  Use patch(PatchContext.of(PatchType.SERVER_SIDE_APPLY), service), or new PatchContext.Builder().withPatchType(PatchType.SERVER_SIDE_APPLY).withForce(true).build() to override conflicts.
 * Fix #2207: added LeaderElector.start to provide a CompletableFuture for easy cancellation.
+* Fix #3758: VersionInfo in KubernetesMockServer can be overridden
 * Fix #3969: relist will not trigger sync events
 * Fix #4082: improving informOnCondition to test the initial list instead of individual add events
 * Fix #3968: SharedIndexInformer.initialState can be used to set the store state before the informer starts.